Extract lightweight cli_util library to fix CMake link error
Summary:
The cli_util_test was failing to link in CMake builds because it linked
against `utils`, which pulls in the entire compression stack (openzl,
csv_parser, parquet, sddl, etc.). The test only uses parse utilities
(checkedstoi, sizeString).

Extract lightweight parsing functions into `parse.h`/`parse.cpp` with
zero heavy dependencies, and create a `cli_util` CMake target for them.
`cli_util_test` now links against `cli_util` only.

Test Plan:
```bash
cmake -B cmakebuild -DOPENZL_BUILD_TESTS=ON
cmake --build cmakebuild --target cli_util_test
./cmakebuild/cli/tests/cli_util_test
```
All 10 GTest cases pass. `make test` also passes (17 Python tests + gtests).
