+++ cli/args/GlobalArgs.h
@@ -5,7 +5,7 @@
 #include <sstream>
 
 #include "cli/cmd.h"
-#include "cli/utils/util.h"
+#include "cli/utils/parse.h"
 
 #include "tools/arg/arg_parser.h"
 #include "tools/arg/parsed_args.h"
